// backend/src/match-series/enums/feedback-option.enum.ts
export enum FeedbackOption {
  TRUST_PLAYER = 'TRUST_PLAYER',
  DEMAND_CARRY = 'DEMAND_CARRY',
  RELIEVE_PRESSURE = 'RELIEVE_PRESSURE',
  DEMAND_AGGRESSION = 'DEMAND_AGGRESSION',
  BLAME_PLAYER = 'BLAME_PLAYER',
  PRAISE_TEAM = 'PRAISE_TEAM',
  REFOCUS_TEAM = 'REFOCUS_TEAM',
  WAKE_UP_TEAM = 'WAKE_UP_TEAM',
  DISAPPOINTED_TEAM = 'DISAPPOINTED_TEAM',
  ABUSIVE_TEAM = 'ABUSIVE_TEAM',
}